extracted logic for handling /set-cookie route into its own file
